weird fix pt2
